[5.7] Allow custom monolog driver to be able to use channel name parser (fallback to environment) and parse string level into monolog constant.
