+++ src/Illuminate/Log/LogManager.php
@@ -18,6 +18,8 @@ use Monolog\Handler\SlackWebhookHandler;
 
 class LogManager implements LoggerInterface
 {
+    use LoggerConfiguration;
+
     /**
      * The application instance.
      *
@@ -38,22 +40,6 @@ class LogManager implements LoggerInterface
      * @var array
      */
     protected $customCreators = [];
-
-    /**
-     * The Log levels.
-     *
-     * @var array
-     */
-    protected $levels = [
-        'debug' => Monolog::DEBUG,
-        'info' => Monolog::INFO,
-        'notice' => Monolog::NOTICE,
-        'warning' => Monolog::WARNING,
-        'error' => Monolog::ERROR,
-        'critical' => Monolog::CRITICAL,
-        'alert' => Monolog::ALERT,
-        'emergency' => Monolog::EMERGENCY,
-    ];
 
     /**
      * Create a new Log manager instance.
@@ -386,40 +372,6 @@ class LogManager implements LoggerInterface
         return tap(new LineFormatter(null, null, true, true), function ($formatter) {
             $formatter->includeStacktraces();
         });
-    }
-
-    /**
-     * Extract the log channel from the given configuration.
-     *
-     * @param  array  $config
-     * @return string
-     */
-    protected function parseChannel(array $config)
-    {
-        if (! isset($config['name'])) {
-            return $this->app->bound('env') ? $this->app->environment() : 'production';
-        }
-
-        return $config['name'];
-    }
-
-    /**
-     * Parse the string level into a Monolog constant.
-     *
-     * @param  array  $config
-     * @return int
-     *
-     * @throws \InvalidArgumentException
-     */
-    protected function level(array $config)
-    {
-        $level = $config['level'] ?? 'debug';
-
-        if (isset($this->levels[$level])) {
-            return $this->levels[$level];
-        }
-
-        throw new InvalidArgumentException('Invalid log level.');
     }
 
     /**
